modified:   README.md 	renamed:    Abstract.md -> docs/Abstract.md 	renamed:    Phase-1.md -> docs/Phase-1.md 	renamed:    Phase-2.md -> docs/Phase-2.md 	renamed:    Phase-3.md -> docs/Phase-3.md 	new file:   docs/Phase-4.md 	modified:   firmware/potentiometer_pid/potentiometer_pid.ino 	new file:   python/ai/__init__.py 	new file:   python/ai/__pycache__/__init__.cpython-311.pyc 	new file:   python/ai/__pycache__/cost_function.cpython-311.pyc 	new file:   python/ai/__pycache__/genetic_tuner.cpython-311.pyc 	new file:   python/ai/cost_function.py 	new file:   python/ai/genetic_tuner.py 	modified:   python/connection_test.py 	new file:   python/genetic_tuner.py 	new file:   python/interface/__init__.py 	new file:   python/interface/__pycache__/__init__.cpython-311.pyc 	new file:   python/interface/__pycache__/motor_interface.cpython-311.pyc 	modified:   python/interface/motor_interface.py 	new file:   python/main_tuner.py 	modified:   python/requirements.txt 	new file:   python/simulation/README.md

+++ firmware/potentiometer_pid/potentiometer_pid.ino
@@ -84,7 +84,7 @@ void loop() {
             Serial.print(",");
             Serial.print(currentPos);
             Serial.print(",");
-            Serial.print(targetSetpoint);
+            Serial.print((int)targetSetpoint);
             Serial.print(",");
             Serial.println(output);
         }
